+++ mafExtractor/src/mafExtractorAPI.h
@@ -31,21 +31,21 @@
 #include "common.h"
 #include "sharedMaf.h"
 
-bool checkRegion(uint32_t targetStart, uint32_t targetStop, uint32_t lineStart, 
-                 uint32_t length, uint32_t sourceLength, char strand);
-bool searchMatched(mafLine_t *ml, const char *seq, uint32_t start, uint32_t stop);
+bool checkRegion(uint64_t targetStart, uint64_t targetStop, uint64_t lineStart, 
+                 uint64_t length, uint64_t sourceLength, char strand);
+bool searchMatched(mafLine_t *ml, const char *seq, uint64_t start, uint64_t stop);
 void printHeader(void);
-uint32_t getTargetColumns(bool **targetColumns, uint32_t *n, mafBlock_t *b, const char *seq, 
-                          uint32_t start, uint32_t stop);
-void printTargetColumns(bool *targetColumns, uint32_t n);
-int32_t **createOffsets(uint32_t n);
-void destroyOffsets(int32_t **offs, uint32_t n);
-mafBlock_t *processBlockForSplice(mafBlock_t *b, uint32_t blockNumber, const char *seq, 
-                                  uint32_t start, uint32_t stop, bool store);
-mafBlock_t *spliceBlock(mafBlock_t *mb, uint32_t l, uint32_t r, int32_t **offsetArray);
-void checkBlock(mafBlock_t *b, uint32_t blockNumber, const char *seq, uint32_t start, 
-                uint32_t stop, bool *printedHeader, bool isSoft);
-void processBody(mafFileApi_t *mfa, char *seq, uint32_t start, uint32_t stop, bool isSoft);
-uint32_t sumBool(bool *array, uint32_t n);
+uint64_t getTargetColumns(bool **targetColumns, uint64_t *n, mafBlock_t *b, const char *seq, 
+                          uint64_t start, uint64_t stop);
+void printTargetColumns(bool *targetColumns, uint64_t n);
+int64_t **createOffsets(uint64_t n);
+void destroyOffsets(int64_t **offs, uint64_t n);
+mafBlock_t *processBlockForSplice(mafBlock_t *b, uint64_t blockNumber, const char *seq, 
+                                  uint64_t start, uint64_t stop, bool store);
+mafBlock_t *spliceBlock(mafBlock_t *mb, uint64_t l, uint64_t r, int64_t **offsetArray);
+void checkBlock(mafBlock_t *b, uint64_t blockNumber, const char *seq, uint64_t start, 
+                uint64_t stop, bool *printedHeader, bool isSoft);
+void processBody(mafFileApi_t *mfa, char *seq, uint64_t start, uint64_t stop, bool isSoft);
+uint64_t sumBool(bool *array, uint64_t n);
 
 #endif // _BLOCK_EXTRACTOR_API_H_
